공지사항 수정 - 글만 수정했을 경우 파일이 삭제되는 것 해결, 수정 페이지에 등록되어있는 파일 표기

+++ koSchool/src/main/java/ko/school/board/controller/NoticeBoardController.java
@@ -137,7 +137,7 @@ public class NoticeBoardController {
 		
 		service.updateNoticBoardService(noticeBoardVo);
 		
-		if(preFileName != null) {
+		if(!file.isEmpty() && preFileName != null) {
 			File file2 = new File(request.getRealPath("/upload") + "/" + preFileName);
 			if(file2.exists()) {
 				file2.delete();
